- create reports page

+++ SQL/procedures.sql
@@ -297,7 +297,7 @@ select id from users where Name=name and EmailId=useremail;
 END$$
 
 DROP PROCEDURE IF EXISTS sp_userupdateprofile;
-CREATE DEFINER=`root`@`localhost` PROCEDURE `sp_userupdateprofile` (`uid` INT(5), `name` VARCHAR(120), `email` VARCHAR(120), `phone` VARCHAR(30), size INT(10))  BEGIN
+CREATE DEFINER=`root`@`localhost` PROCEDURE `sp_userupdateprofile` (`uid` INT(5), `name` VARCHAR(120), `email` VARCHAR(120), `phone` VARCHAR(30), size DECIMAL(15,2))  BEGIN
 update users set Name=name, LastUpdationDate=current_timestamp(), EmailId=email, PhoneNumber=phone, Size=size WHERE id=uid;
 END$$
